+++ CHANGELOG.md
@@ -15,16 +15,19 @@
 
 * Removing Accounts implemented with full cleanup of associated data
   * If admin account, correctly removes from the ADMIN list
+  * Can not remove the Mediator DID or Root-Admin DID
 * Database Schema version is now recorded, allows for upgrade paths when schema changes
 * Mediator Account Type added, allows for treating the Mediator DID separately
 * FIX: Trying to strip admin rights from an empty list will now correctly create a ProblemReport that explains the issue
 * FIX: Mediator Administration generates a client side error when no Admin DID is selected when removing Admin Accounts
 * FIX: Double hashing of DID's on admin_add, refactored so now only uses SHA256 hashed DID's
+* FEATURE: Added AccountChangeType to the mediator account-management protocol
 
 ### SDK (0.9.4)
 
 * FIX: If ProblemReport had no args, deserializing would fail as no args field. Now defaults to empty array correctly
 * TEST: Added ProblemReport tests to check for empty args and serialization/deserialization
+* FEATURE: Added AccountChangeType to the mediator account-management protocol
 
 ## 18th January 2025 (0.9.2)